== misc == *Some misc optimisation. -Please don't forget to add the upgrade_15-3athena_svn477.sql file!
*Some item updates.
-Fixed some item bonusses.
-Added some missing items.

*ALL_REVERSEORCISH
-Added full support for this skill.

*A little cleanup for Pets
-Separated damage value for petskillattack2 to its own var 'damage', that previously use 'lv' var to store the value.
-Added validation for pet script commands
--Skill for petskillattack, petskillattack2, & petskillsupport
--SC range for petrecovery.

*Fixed another potential map crash.
-SC_EQC & SC_CBC shouldn't be saved on logout.

+++ src/map/unit.h
@@ -109,7 +109,7 @@ int unit_warp(struct block_list *bl, short map, short x, short y, clr_type type)
 int unit_setdir(struct block_list *bl,unsigned char dir);
 uint8 unit_getdir(struct block_list *bl);
 int unit_blown(struct block_list* bl, int dx, int dy, int count, int flag);
-int unit_blown_immune(struct block_list* bl, int flag);
+uint8 unit_blown_immune(struct block_list* bl, uint8 flag);
 
 // ‚»‚±‚Ü‚Å•às‚Å‚½‚Ç‚è’…‚¯‚é‚©‚Ì”»’è
 bool unit_can_reach_pos(struct block_list *bl,int x,int y,int easy);
